requirements.txt
hyperliquid-python-sdk
python-dotenv

